complete data.aggregate, remove aggregate.js
