--- src/compiler/aggregate.js
@@ -1,39 +1,0 @@
-'use strict';
-
-require('../globals');
-
-var util = require('../util');
-
-module.exports = aggregates;
-
-function aggregates(dataTable, encoding, opt) {
-  opt = opt || {};
-
-  var dims = {}, meas = {};
-
-  encoding.forEach(function(field, encType) {
-    if (field.aggregate) {
-      if (field.aggregate === 'count') {
-        meas.count = {op: 'count', field: '*'};
-      }else {
-        meas[field.aggregate + '|'+ field.name] = {
-          op: field.aggregate,
-          field: encoding.fieldRef(encType, {nofn: true})
-        };
-      }
-    } else {
-      dims[field.name] = encoding.fieldRef(encType);
-    }
-  });
-  dims = util.vals(dims);
-  meas = util.vals(meas);
-
-  if (meas.length > 0) {
-    if (!dataTable.transform) dataTable.transform = [];
-    dataTable.transform.push({
-      type: 'aggregate',
-      groupby: dims,
-      fields: meas
-    });
-  }
-}
